Moving all loggers into lib/log/
