+++ src/bin/download.ts
@@ -3,7 +3,7 @@ import DataDir from '../lib/data/dir';
 import { DataSet } from '../lib/data/set';
 import { downloadAllData } from '../lib/engine/download';
 import { Hubspot } from '../lib/hubspot';
-import { Logger } from '../lib/log/logger';
+import { Logger } from '../lib/log';
 
 const dataDir = DataDir.root.subdir("in");
 const log = new Logger(dataDir.subdir("log"));
