+++ src/lib/hubspot/api.ts
@@ -1,6 +1,6 @@
 import * as hubspot from '@hubspot/api-client';
 import assert from 'assert';
-import { Logger } from '../log/logger';
+import { Logger } from '../log';
 import { hubspotCredsFromENV } from '../parameters/env-config';
 import { KnownError } from '../util/errors';
 import { batchesOf, isPresent } from '../util/helpers';
