+++ src/lib/marketplace/structure.ts
@@ -1,4 +1,4 @@
-import { Logger } from "../log/logger";
+import { Logger } from "../log";
 import { Table } from "../log/table";
 import { formatMoney } from "../util/formatters";
 import { isPresent } from "../util/helpers";
